+++ backend/src/errors/mod.rs
@@ -1,0 +1,2 @@
+pub mod app_error;
+pub mod validation;
