+++ backend/src/routes/parts.rs
@@ -1,172 +1,13 @@
-use axum::response::{IntoResponse, Response};
-use axum::{Json, extract::Path, extract::State, http::StatusCode};
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+use crate::errors::app_error::AppError;
+use crate::errors::validation::extract_validation_errors;
+use crate::models::part::{NewPart, Part};
+use crate::responses::success::SuccessResponse;
+
+use axum::{Json, extract::Path, extract::State};
 use sqlx::PgPool;
 use tracing::{error, info};
 use uuid::Uuid;
-use validator::{Validate, ValidationErrors};
-
-#[derive(Debug, Serialize)]
-pub struct ErrorResponse {
-    pub success: bool,
-    pub code: u16,
-    pub error: ErrorDetail,
-}
-
-#[derive(Debug, Serialize)]
-pub struct ErrorDetail {
-    pub message: String,
-}
-
-#[derive(Debug)]
-pub enum AppError {
-    ValidationError(ValidationErrorResponse),
-    NotFound(String),
-    DatabaseError(String),
-    // InternalError(String),
-}
-
-impl IntoResponse for AppError {
-    fn into_response(self) -> Response {
-        let response: Response = match self {
-            AppError::ValidationError(errors) => {
-                let status = StatusCode::BAD_REQUEST;
-
-                error!(
-                    "Validation error ({}): {} error(s) - {:?}",
-                    status,
-                    errors.errors.len(),
-                    errors.errors
-                );
-
-                let body = Json(errors);
-
-                (status, body).into_response()
-            }
-            AppError::NotFound(message) => {
-                let status = StatusCode::NOT_FOUND;
-
-                error!("NotFound error ({}): {}", status, message);
-
-                let body = Json(ErrorResponse {
-                    success: false,
-                    code: status.as_u16(),
-                    error: ErrorDetail { message },
-                });
-
-                (status, body).into_response()
-            }
-            AppError::DatabaseError(message) => {
-                let status = StatusCode::INTERNAL_SERVER_ERROR;
-
-                error!("Database error ({}): {}", status, message);
-
-                let body = Json(ErrorResponse {
-                    success: false,
-                    code: status.as_u16(),
-                    error: ErrorDetail { message },
-                });
-
-                (status, body).into_response()
-            } // AppError::InternalError(msg) => (StatusCode::INTERNAL_SERVER_ERROR, msg),
-        };
-        response
-    }
-}
-
-#[derive(Serialize, Debug)]
-pub struct ValidationErrorResponse {
-    pub success: bool,
-    pub code: u16,
-    pub errors: Vec<FieldError>,
-}
-
-#[derive(Serialize, Debug)]
-pub struct FieldError {
-    pub field: String,
-    pub message: String,
-}
-
-fn extract_validation_errors(errors: ValidationErrors) -> ValidationErrorResponse {
-    let mut field_errors = Vec::new();
-
-    for (field, errors) in errors.field_errors() {
-        for error in errors {
-            let message = error
-                .message
-                .clone()
-                .unwrap_or_else(|| std::borrow::Cow::from("Invalid input"));
-
-            field_errors.push(FieldError {
-                field: field.to_string(),
-                message: message.to_string(),
-            });
-        }
-    }
-    ValidationErrorResponse {
-        success: false,
-        code: StatusCode::BAD_REQUEST.as_u16(),
-        errors: field_errors,
-    }
-}
-
-#[derive(Debug, Serialize)]
-pub struct SuccessResponse<T> {
-    pub success: bool,
-    pub code: u16,
-    pub data: T,
-}
-
-impl<T> SuccessResponse<T> {
-    pub fn ok(data: T) -> Self {
-        SuccessResponse {
-            success: true,
-            code: StatusCode::OK.as_u16(),
-            data,
-        }
-    }
-
-    pub fn created(data: T) -> Self {
-        SuccessResponse {
-            success: true,
-            code: StatusCode::CREATED.as_u16(),
-            data,
-        }
-    }
-
-    pub fn no_content() -> Self
-    where
-        T: Default,
-    {
-        SuccessResponse {
-            success: true,
-            code: StatusCode::NO_CONTENT.as_u16(),
-            data: T::default(),
-        }
-    }
-}
-
-#[derive(sqlx::FromRow, Serialize)]
-pub struct Part {
-    pub id: Uuid,
-    pub part_number: String,
-    pub name: String,
-    pub description: Option<String>,
-    pub kind: Option<String>,
-    pub created_at: Option<DateTime<Utc>>,
-    pub updated_at: Option<DateTime<Utc>>,
-}
-
-#[derive(Deserialize, Validate)]
-pub struct NewPart {
-    #[validate(length(min = 1, message = "part_number must not be empty"))]
-    pub part_number: String,
-    #[validate(length(min = 1, message = "name must not be empty"))]
-    pub name: String,
-    pub description: Option<String>,
-    pub kind: Option<String>,
-}
+use validator::Validate;
 
 // #[axum::debug_handler]
 pub async fn create_part(
